requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Django==3.2.9
django-croppie==1.3
django-mapbox-location-field==1.7.3
django-taggit==2.0.0
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
geos==0.2.3
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
outcome==1.1.0
Pillow==8.4.0
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
